refactor ignore command to typescript
