--- lib/__mocks__/utils.js
@@ -1,6 +1,0 @@
-jest.unmock('../utils')
-const utils = require('../utils')
-
-utils.pushTempRepo = jest.fn()
-
-module.exports = utils
