--- lib/constants.js
@@ -1,12 +1,0 @@
-const os = require('os')
-const path = require('path')
-
-const CONFIG_FILENAME = 'git-slice.json'
-
-const INVALID_ARG_MSG =
-  'Invalid arguments, following are the usage details of this command:'
-
-module.exports = {
-  CONFIG_FILENAME,
-  INVALID_ARG_MSG,
-}
