+++ src/cmds/ignore.ts
@@ -1,24 +1,32 @@
-const path = require('path')
-const Git = require('nodegit')
-const fs = require('fs-extra')
-const _ = require('lodash')
-const { CONFIG_FILENAME } = require('../constants')
+import Git from 'nodegit';
+import path from 'path';
+import fs from 'fs-extra';
+import _ from 'lodash';
 
-async function modifyIgnoredFiles(currentDir, filesToAdd, filesToRemove) {
+import { CONFIG_FILENAME } from '@lib/constants';
+import { GitSliceConfigType } from '@customTypes/graphql';
+
+export default async (currentDir: string, filesToAdd: string[], filesToRemove: string[]) => {
   try {
-    const config = await fs.readJson(path.resolve(currentDir, CONFIG_FILENAME))
+    const config: GitSliceConfigType = await fs.readJson(path.resolve(currentDir, CONFIG_FILENAME))
+
     const folderRepo = await Git.Repository.open(currentDir)
     await folderRepo.checkoutBranch('master', new Git.CheckoutOptions())
+
     const commonFileExists = filesToAdd.some(function(v) {
       return filesToRemove.indexOf(v) >= 0
     })
+
     if (commonFileExists) {
       throw 'Error: Both add and remove operation is being performed on the same file'
     }
+
     let updatedIgnoredFiles = config.ignore
+
     if (filesToAdd.length) {
       updatedIgnoredFiles = _.uniq([...updatedIgnoredFiles, ...filesToAdd])
     }
+    
     if (filesToRemove.length) {
       updatedIgnoredFiles = updatedIgnoredFiles.filter(
         x => filesToRemove.indexOf(x) === -1
@@ -54,5 +62,3 @@ async function modifyIgnoredFiles(currentDir, filesToAdd, filesToRemove) {
     return Promise.reject(e)
   }
 }
-
-module.exports = modifyIgnoredFiles
